chore(package): renamed server types to interfaces
