+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './serverTypes';
+export * from './interfaces/AppleMusic';
 export * from './appleMusicError';
 export * from './calendarDate';
 export * from './client';
