+++ src/interfaces/AppleMusic/album.ts
