+++ src/interfaces/AppleMusic/albumRelationship.ts
