+++ src/interfaces/AppleMusic/artist.ts
