+++ src/interfaces/AppleMusic/artistRelationship.ts
