+++ src/interfaces/AppleMusic/artistResponse.ts
