+++ src/interfaces/AppleMusic/artwork.ts
