+++ src/interfaces/AppleMusic/contentRating.ts
