+++ src/interfaces/AppleMusic/curator.ts
