+++ src/interfaces/AppleMusic/curatorRelationship.ts
