+++ src/interfaces/AppleMusic/error.ts
