+++ src/interfaces/AppleMusic/genre.ts
