+++ src/interfaces/AppleMusic/genreRelationship.ts
