+++ src/interfaces/AppleMusic/index.ts
