+++ src/interfaces/AppleMusic/musicVideo.ts
