+++ src/interfaces/AppleMusic/musicVideoRelationship.ts
