+++ src/interfaces/AppleMusic/musicVideoResponse.ts
