+++ src/interfaces/AppleMusic/playlist.ts
