+++ src/interfaces/AppleMusic/preview.ts
