+++ src/interfaces/AppleMusic/relationship.ts
