+++ src/interfaces/AppleMusic/resource.ts
