+++ src/interfaces/AppleMusic/searchResponse.ts
