+++ src/interfaces/AppleMusic/song.ts
