+++ src/interfaces/AppleMusic/songResponse.ts
