+++ src/interfaces/AppleMusic/station.ts
