+++ src/interfaces/AppleMusic/stationResponse.ts
