+++ src/interfaces/AppleMusic/trackRelationship.ts
